--- router/main.go
@@ -1,16 +1,0 @@
-package main
-
-import (
-	"SimpleDouyin/service"
-	"github.com/gin-gonic/gin"
-)
-
-func main() {
-	go service.RunMessageServer()
-
-	r := gin.Default()
-
-	initRouter(r)
-
-	r.Run() // listen and serve on 0.0.0.0:8080 (for windows "localhost:8080")
-}
